perf: Replace Puppeteer with html-to-image for screenshot generation
This commit introduces the html-to-image library for generating screenshots. Previously, we were using an API endpoint that returned a base64 image. This endpoint was an Express server that used Puppeteer for screenshot generation.

However, this method was slow and inefficient. By switching to the client-side html-to-image library, we've significantly improved performance.